replace phonimizer with g2p

+++ egs/ljspeech/tts/vits/generator.py
@@ -515,10 +515,12 @@ class VITSGenerator(torch.nn.Module):
         cum_dur_flat = cum_dur.view(b * t_x)
         path = torch.arange(t_y, dtype=dur.dtype, device=dur.device)
         path = path.unsqueeze(0) < cum_dur_flat.unsqueeze(1)
-        path = path.view(b, t_x, t_y).to(dtype=mask.dtype)
+        # path = path.view(b, t_x, t_y).to(dtype=mask.dtype)
+        path = path.view(b, t_x, t_y).to(dtype=torch.float)
         # path will be like (t_x = 3, t_y = 5):
         # [[[1., 1., 0., 0., 0.],      [[[1., 1., 0., 0., 0.],
         #   [1., 1., 1., 1., 0.],  -->   [0., 0., 1., 1., 0.],
         #   [1., 1., 1., 1., 1.]]]       [0., 0., 0., 0., 1.]]]
         path = path - F.pad(path, [0, 0, 1, 0, 0, 0])[:, :-1]
+        # path = path.to(dtype=mask.dtype)
         return path.unsqueeze(1).transpose(2, 3) * mask
